Move `is_a_lambda` to TypeTraits
